+++ server/app.js
@@ -20,7 +20,7 @@ const port = 3001;
 
 const articleRouter = require('./router/articles');
 const userRouter = require('./router/users');
-/* const web3Router = require('./router/web3'); */
+const web3Router = require('./router/web3');
 
 app.use(cors({
   origin: true,
@@ -46,7 +46,7 @@ app.use(cookieParser())
 app.use('/articles', articleRouter);
 app.use('/users', userRouter);
 
-//app.use('/', web3Router);
+app.use('/web3', web3Router);
 
 
 app.get('/', (req, res) => {
