+++ server/router/web3.js
@@ -1,8 +1,7 @@
-const { reward, transfer } = require('../controller/web3');
+const { userSend } = require('../controller/web3');
 const express = require('express');
 const router = express.Router();
 
-router.post('/reward', reward);
-router.post('/transfer', transfer);
+router.post('/userSend', userSend);
 
 module.exports = router;
